This is a major change to the proxy. The proxy should now handle filters (not fully tested yet) for the headers and body of any request or response.
Whether a particular filter is applied or not depends on a
match subroutine (a closure actually) that is compiled by
the _push_filter() method.

New subroutines:

_push_filter: main code for adding filters
push_headers_filter: wrapper around _push_filter
push_body_filter: wrapper around _push_filter
filter_headers: a fonction to call the filters on a message header
filter_body: a function to call the filters on a message body
_proxy_headers_filter: default headers filter for HTTP::Proxy (always applied)

A new (uncomplete) test file is provided too.