The proxy now supports persistent connections!! Added the maxserve attribute.
Most of the code to support chunked transfer encoding is taken from
HTTP::Daemon. The reason I did not directly use the send_response
method is that the form that requires a callback cannot be called within
the LWP::UserAgent simple_request method, in the form that accepts
a callback.