A new class HTTP::Proxy::FilterStack to handle the filters and buffer information left over by a proxy for the next pass.
(This can be used for example if a filter needs to handle data
line by line: one could stack a filter that make sure only
full lines are passed to the next filters in tha stack, and would
keep the rest of the line in the given buffer).

Next step is to handle nochunks filters, and/or add a HTTP::Proxy::Filter
class.