There was a big bug in the test suite for the keep-alive. The goal was to have a proxy that processes no more than 3 requests per connection handle 1 request by a client than does not support keep-alive, and 6 requests by a client that supports keep alive.
But mu first version of the client was simply a LWP::UA that
supports keep-alive but lied and sent a Connection: close header.
So the proxy did close the connection, but the client (doing
some keep-alive) kept (on some machines) sending the next request
through the first connection. Which was close by the proxy.
I suppose this happened when the client didn't detect that the
proxy closed the connection on its side (slow machines?).

Anyway, keepalive tests should work better now.