remove the Accept-Encoding header as soon as there is a body filter
The issue with the previous implementation is that the "current"
list of body filters has not been computed yet when the standard
header filter kicks in. So the Accept-Encoding was NEVER removed.

With this patch, the removal will happen as soon as the proxy has
a body filter, even if it's not going to be used for this specific
request/response. A possible improvement would be to precompute
a list of "likely" body filters (only the mime condition cannot
be computed at request time), and then filter that list again
when the response actually arrives.